Stylized with the following command: astyle --indent=spaces=2  --convert-tabs -bLS --mode=c `find . -type f -name '*.cc'`
This command tries to mimic the manual formatted style used so far.
Checking the files, the only inconsistency I have found is the multilines method calls, Now being aligned to the parenthesis of the method. What is very sensible anyway.

I have found that in the rest of the files it made a good job (more readable) or simply a non-noticeable job.

I tried to run it on the .hh files but I don't know how to avoid the stylizer change the labels (make all the public and private in their own lines)